requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiolimiter==1.2.1
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asgiref==3.8.1
assemblyai==0.40.2
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
baidusearch==1.0.3
beartype==0.12.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
boto3==1.37.35
botocore==1.37.35
browser-use==0.1.40
browsergym==0.13.3
browsergym-assistantbench==0.13.3
browsergym-core==0.13.3
browsergym-experiments==0.13.3
browsergym-miniwob==0.13.3
browsergym-visualwebarena==0.13.3
browsergym-webarena==0.13.3
browsergym-workarena==0.4.1
cachetools==5.5.1
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
dataclasses-json==0.6.7
datasets==3.4.1
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.9
distro==1.9.0

django-mssql-backend==2.8.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
dnspython==2.7.0
docker==7.1.0
duckduckgo_search==7.5.5
elevenlabs==0.2.26
email-validator==2.0.0
english-words==2.0.1
et_xmlfile==2.0.0
evaluate==0.4.3
executing==2.2.0
Faker==37.1.0
Farama-Notifications==0.0.4
fastapi==0.115.12
filelock==3.18.0
Flask==2.3.3
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.5.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.12.1
google-generativeai==0.8.5
googleapis-common-protos==1.66.0
googlesearch-python==1.3.0
greenlet==3.2.0
grpcio==1.70.0
grpcio-status==1.70.0
gymnasium==1.1.1
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==0.3.3
langchain-core==0.3.52
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.31
lazy_loader==0.4
libvisualwebarena==0.0.15
libwebarena==0.0.4
loguru==0.7.3
lxml==5.3.2
Mako==1.3.9
markdownify==0.14.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.5.0
monotonic==1.6
mpmath==1.3.0
mssql-django==1.5
multidict==6.4.3
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
ollama==0.4.7

openpyxl==3.1.5
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.8
playwright==1.51.0
pluggy==1.5.0
portalocker==3.1.1
posthog==3.25.0
primp==0.14.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.0
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyee==12.1.1
pygame==2.6.1
Pygments==2.19.1
pyodbc==5.2.0
pyparsing==3.2.1
pytest==8.3.5
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.11.4
sacrebleu==2.5.1
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
setuptools==75.8.2
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.23
sqlparse==0.5.3
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
text-generation==0.7.0
tifffile==2025.3.30
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
types-requests==2.32.0.20250328
types-tqdm==4.67.0.20250404
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
unidiff==0.7.5
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.1
virtualenv==20.31.2
wcwidth==0.2.13
weblinx==0.3.2
weblinx-browsergym==0.0.2
websockets==15.0.1
Werkzeug==2.3.7
wsproto==1.2.0
WTForms==3.2.1
xxhash==3.5.0
yarl==1.19.0
zstandard==0.23.0
